services/ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Event-driven ingestion worker (EDA) for RAG content and product cards"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.34",
  "certifi>=2024.0",
  "openai>=1.0",
  "pydantic>=2.6",
  "weaviate-client>=4.0",
  "httpx>=0.27",
  "yt-dlp>=2024.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
